Add my content to template.
Removed scripts that generate the dist file. On relaunching local host for dev they largely undo a lot if the styling.
Also Note: the set of files outside of the dif are the ones currently displayed on git hub, has not yet been configured to look for the html in the dif.